Added a sorting function in main.js
I made a sorting function, I have no idea how the targetData is served to the function. What I am guessing is served is the list of items defined at the top of the main.js file. 
The function modifies the targetData array, and then returns it.

If I didn't do it right or you want me to make changes, message me on discord on @CamoBrie#2115